requirements.txt
interception-python
XInput-Python
numpy
scipy
pyautogui
ctypes